Reorganize AssemblerTraverser into traversal
