+++ lib/labelhash.hh
@@ -44,7 +44,6 @@ Contact: khmer-project@idyll.org
 #include <string>
 #include <utility>
 
-#include "hashbits.hh"
 #include "hashtable.hh"
 #include "khmer.hh"
 #include "read_parsers.hh"
@@ -193,16 +192,6 @@ public:
     void label_across_high_degree_nodes(const char * sequence,
                                         SeenSet& high_degree_nodes,
                                         const Label label);
-
-    /*
-    std::vector<std::string> assemble_labeled_path(const Kmer seed_kmer) const;
-    void _assemble_labeled_right(const char * start_kmer,
-                                 std::vector<std::string>& paths,
-                                 SeenSet& visited) const;
-    std::string _assemble_linear_labels(const std::string kmer,
-                                        const Label label,
-                                        SeenSet& visited) const;
-    */
 };
 }
